requirements.txt
absl-py==2.2.2
addict==2.4.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==23.2.0
boto3==1.34.34
botocore==1.34.162
branca==0.8.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
cog==0.14.12
colorama==0.4.6
crispy-bootstrap5==2024.2
cryptography==45.0.2
defusedxml==0.7.1
distro==1.9.0
Django==5.0.2
django-allauth==0.61.1
django-cleanup==8.0.0
django-crispy-forms==2.1
django-debug-toolbar==4.3.0
django-environ==0.11.2
django-storages==1.14.2
Faker==37.3.0
fastapi==0.115.12
filelock==3.18.0
folium==0.19.6
fsspec==2025.3.2
future==1.0.0
grpcio==1.71.0
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
imageio==2.37.0
inflection==0.5.1
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
lazy_loader==0.4
lmdb==1.6.2
Markdown==3.8
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
oauthlib==3.2.2
openai==1.79.0
openapi==2.0.0
opencv-contrib-python==4.11.0.86
opencv-python==4.9.0.80
packaging==25.0
pandas==2.2.3
pillow==10.2.0
platformdirs==4.3.8
protobuf==6.31.0
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.10.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.31.0
requests-oauthlib==2.0.0
rpds-py==0.25.0
s3transfer==0.10.4
scikit-image==0.25.2
scipy==1.15.3
setuptools==80.7.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.46.2
structlog==24.4.0
sympy==1.14.0
tb-nightly==2.20.0a20250518
tensorboard-data-server==0.7.2
tifffile==2025.5.10
torch==2.2.1
torchvision==0.17.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
watchfiles==1.0.5
websockets==15.0.1
Werkzeug==3.1.3
whitenoise==6.6.0
xyzservices==2025.4.0
yapf==0.43.0
python-decouple==3.8
django-cors-headers==4.3.1
